Mar 1, 2025, 6:32 AM

+++ CT06_07/lesson7.py
@@ -17,6 +17,6 @@ average_score = total / 3
 
 student_name = "Alex"
 
-average_score = float(average_score)
+average_score = int(average_score)
 
 print("Average score for " + student_name + " is: " + str(average_score))
